Extract contexts
